requirements.txt
google-cloud-vision
Django==3.1.1
